Move files in better structure
